Reconfigured the uworker packer.io build for openstack so that it works with openstack auth v3.
Summary:
Before SMOG broke down, they used openstack auth v2. V2 is now deprecated and not supported. Packer.io have a 'fix' for this. It is described in the ducumentation of packer builders https://www.packer.io/docs/builders/openstack.html.

Also fixed some minor changes in the setup to make it work with the recent versions of u-service.

Note that a file with the setting and credentials to the u-service is not in this repo (and shouldn't be) it looks like this:

  export UWORKER_JOB_API_ROOT=http://odin.rss.chalmers.se:8080/rest_api/
  export UWORKER_JOB_API_USERNAME=user
  export UWORKER_JOB_API_PASSWORD=password
  export UWORKER_REGISTRY_URL=https://docker2.molflow.com
  export UWORKER_REGISTRY_USERNAME=user
  export UWORKER_REGISTRY_PASSWORD=password

Test Plan: There are no test for this. We can verify that it works by going to http://odin.rss.chalmers.se:8080/MESO21ALL to check that the processing is running and returning results.

Reviewers: skymandr, bengt.rydberg, gregoire

Reviewed By: gregoire

Subscribers: jenkins

Maniphest Tasks: T546

Differential Revision: https://phabricator.molflow.com/D318